// frontend/src/types/gen/service/PerformanceListResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ServicePerformancePageItem } from './PerformancePageItem.ts'
import type { ServicePerformanceStatsData } from './PerformanceStatsData.ts'

export type ServicePerformanceListResponse = {
  /**
   * @type array | undefined
   */
  list?: ServicePerformancePageItem[]
  /**
   * @type object | undefined
   */
  stats?: ServicePerformanceStatsData
  /**
   * @type integer | undefined
   */
  total?: number
}