// frontend/src/types/gen/service/ResourcePerformanceListResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ServiceResourcePerformanceItem } from './ResourcePerformanceItem.ts'

export type ServiceResourcePerformanceListResponse = {
  /**
   * @type array | undefined
   */
  list?: ServiceResourcePerformanceItem[]
  /**
   * @type integer | undefined
   */
  total?: number
}